android: Streamline the creation of an Android application
First android specific class in the Pris backend: ClutterAndroidApplication

This class is used to abstract the initialization and the life cycle of
an android application. This commit only implements the initialization
phase and we happily close our eyes on other events.

The test-actors android example has been update to use the new API.